added relation betw tables & some functionalities

+++ app/Http/Controllers/Admin/PostController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Post;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -21,6 +22,13 @@ class PostController extends Controller
                 'max:100'
             ]
         ];
+    }
+
+
+    public function userIndex() {
+        $posts = Post::where('user_id', Auth::user()->id)->paginate(10);
+
+        return view('admin.posts.index', compact('posts'));
     }
 
     /**
@@ -55,7 +63,10 @@ class PostController extends Controller
     {
         $request->validate($this->confValidation(null));
 
-        $post = Post::create($request->all());
+        $data = $request->all() + [
+            'user_id' => Auth::user()->id
+        ];
+        $post = Post::create($data);
 
         return redirect()->route('admin.posts.show', $post->slug);
     }
@@ -79,6 +90,8 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if (Auth::user()->id !== $post->user_id) abort(403);
+
         return view('admin.posts.edit', compact('post'));
     }
 
@@ -91,6 +104,8 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        if (Auth::user()->id !== $post->user_id) abort(403);
+
         $request->validate($this->confValidation($post));
 
         $post->update($request->all());
@@ -106,6 +121,8 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if (Auth::user()->id !== $post->user_id) abort(403);
+
         $post->delete();
 
         return redirect()->route('admin.posts.index');
